Cleaner port checking code.
Removed redundant code in NewTinyServer / NewWebServer constructors.

Added more examples of routing. It is possible to specify routes to be
more specific even if they are sharing same URL scheme.

Example: When the routes are defined in this particular order
        /page/about
        /page/<page_id>

The handler for /page/about is invoked when the url is exactly
specified. If the order were to be reversed, then the first match would
be the most generic one and its handler is called.

Consolidated utilities into util.go