TinyServer.Start() is now non-blocking. It can be stopped by calling server.Stop(), which essentially closes a private channel that we are listening for messages in a select in Accept() method.
Since the default Accept() is blocking, we set a
timeout on it and then verify if channel is still open
which if it is, we call Accept again.

The drawback is if there are multiple conections
within the span of the timeout, this can lead to
unknown behaviour.

This solution is more of a hack to get this working.